warning
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@13016 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Graphics/BIOS/biossums.c
@@ -146,7 +146,7 @@ int main(int argc, char* argv[])
 
   if (modified == 1) {
 #ifdef VBOX
-    long new_bios_len;
+    size_t new_bios_len;
 #endif
     if ((stream = fopen( argv[1], "wb")) == NULL) {
       printf("Error opening %s for writing.\n", argv[1]);
